DPC-745: Moved data factory into its own package.
Created a factories package and renamed the generate class into a more specific one.
