+++ dpc-common/src/test/java/gov/cms/dpc/fhir/validations/EndpointValidationTest.java
@@ -4,21 +4,17 @@ import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.validation.FhirValidator;
 import ca.uhn.fhir.validation.ValidationOptions;
 import ca.uhn.fhir.validation.ValidationResult;
-import gov.cms.dpc.fhir.validations.profiles.AddressProfile;
 import gov.cms.dpc.fhir.validations.profiles.EndpointProfile;
 import gov.cms.dpc.testing.BufferedLoggerHandler;
 import org.hl7.fhir.dstu3.hapi.ctx.DefaultProfileValidationSupport;
 import org.hl7.fhir.dstu3.hapi.validation.FhirInstanceValidator;
 import org.hl7.fhir.dstu3.hapi.validation.ValidationSupportChain;
 import org.hl7.fhir.dstu3.model.*;
-import org.hl7.fhir.dstu3.model.codesystems.EndpointConnectionType;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.util.List;
-
-import static gov.cms.dpc.testing.DataFactories.createFakeEndpoint;
+import static gov.cms.dpc.testing.factories.OrganizationFactory.createFakeEndpoint;
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(BufferedLoggerHandler.class)
