+++ dpc-common/src/test/java/gov/cms/dpc/fhir/validations/OrganizationValidationTest.java
@@ -5,25 +5,21 @@ import ca.uhn.fhir.validation.FhirValidator;
 import ca.uhn.fhir.validation.ValidationOptions;
 import ca.uhn.fhir.validation.ValidationResult;
 import gov.cms.dpc.fhir.DPCIdentifierSystem;
-import gov.cms.dpc.fhir.validations.profiles.AddressProfile;
-import gov.cms.dpc.fhir.validations.profiles.EndpointProfile;
 import gov.cms.dpc.fhir.validations.profiles.OrganizationProfile;
 import gov.cms.dpc.testing.BufferedLoggerHandler;
 import org.hl7.fhir.dstu3.hapi.ctx.DefaultProfileValidationSupport;
 import org.hl7.fhir.dstu3.hapi.validation.FhirInstanceValidator;
 import org.hl7.fhir.dstu3.hapi.validation.ValidationSupportChain;
 import org.hl7.fhir.dstu3.model.Address;
-import org.hl7.fhir.dstu3.model.Meta;
 import org.hl7.fhir.dstu3.model.Organization;
-import org.hl7.fhir.dstu3.model.Reference;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.Collections;
 
-import static gov.cms.dpc.testing.DataFactories.generateFakeAddress;
-import static gov.cms.dpc.testing.DataFactories.generateFakeOrganization;
+import static gov.cms.dpc.testing.factories.OrganizationFactory.generateFakeAddress;
+import static gov.cms.dpc.testing.factories.OrganizationFactory.generateFakeOrganization;
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(BufferedLoggerHandler.class)
